[gNMI-1.14] prune device config before push (#3573)
* prune baseconfig

* removed unused

* fix readme

* fix readme

* fix readme

---------
