+++ feature/system/gnmi/metadata/tests/large_set_consistency_test/README.md
@@ -40,3 +40,13 @@ dut.testbed
 
 ## Minimum DUT platform
 FFF
+
+## OpenConfig Path and RPC Coverage
+
+```yaml
+rpcs:
+  gnmi:
+    gNMI.Get:
+    gNMI.Subscribe:
+
+```
