+++ feature/system/gnmi/metadata/tests/large_set_consistency_test/metadata.textproto
@@ -13,4 +13,12 @@ platform_exceptions: {
     default_network_instance: "default"
   }
 }
+platform_exceptions: {
+  platform: {
+    vendor: CISCO
+  }
+  deviations: {
+    skip_macaddress_check: true
+  }
+}
